refactor(tests): move files into correct folder
