+++ app/tests/api/v1/endpoints/test_server.py
@@ -2,24 +2,24 @@ from typing import Tuple
 
 from fastapi.responses import Response
 
-from app import schemas
+from app import crud, schemas
 from app.models import Server
 from app.tests.test_base import TestBase
 
 
 class TestServerRoute(TestBase):
-    def _create_server(self, url: str = "http://testserver") -> Tuple[Response, int]:
-        initial_count = len(Server.get_all(self.db))
+    def _create_server(self, url: str = "http://testserver2") -> Tuple[Response, int]:
+        initial_count = len(crud.server.get_multi(self.db))
         data = schemas.ServerCreate(
             url=url,
             client_id="some_client_id",
             client_secret="some_client_secret",
-        ).dict()
-        response = self.client.post("/api/v1/servers", json=data)
+        ).dict(exclude_unset=True)
+        response = self.client.post("/api/v1/servers/", json=data)
         return response, initial_count
 
     def _cleanup_server(self):
-        self.db.query(Server).filter(Server.url == "http://testserver").delete()
+        self.db.query(Server).filter(Server.url == "http://testserver2").delete()
         self.db.commit()
 
     def test_bad_url_rejected(self):
@@ -33,14 +33,14 @@ class TestServerRoute(TestBase):
         expected_keys = ["id", "url"]
         assert response.status_code == 201
         assert expected_keys == list(response.json().keys())
-        assert len(Server.get_all(self.db)) == initial_count + 1
+        assert len(crud.server.get_multi(self.db)) == initial_count + 1
 
         # Test trying to create a different server with the same URL
         # returns a 400 response
         response, _ = self._create_server()
         assert response.status_code == 400
         assert response.json() == {
-            "detail": "Server with url 'http://testserver' already exists."
+            "detail": "Server http://testserver2 already configured."
         }
         self._cleanup_server()
